move defs to cellnet
